// src/types/Product.ts
  export interface Volume {
    id: string
    label: string
    in_stock: boolean
  }
  
  export interface Product {
    id: string
    slug: string
    name: string
    category: string
    image: string
    price: number
    old_price?: number
    discount_percent: number
    currency: string
    rating: number
    reviews_count: number
    in_stock: boolean
    selected_volume_id: string
    volumes: Volume[]
  }

  export interface ProductsResponse {
    success: boolean
    data: {
      products: Product[]
    }
  }